functions/requirements.txt
firebase_functions~=0.5.0
firebase_admin
openai>=1.75.0
line-bot-sdk~=3.22.0